refactor: update path structure
